02 13 2025
I had a great time building up to solve this query.

I have learned a new way of building up which inspires confidence in the path.  and that is to slowly and visibly build up and get result sets that move in the general direction that I believe the query wants.